@@ -25,5 +25,18 @@ class UsersController < ApplicationController
       redirect_to welcome_path, :notice => 'You are not logged in.'
     end
   end
-  
+
+  private
+
+  def user_params
+    if current_user.admin
+      params.require(:person).permit(
+        :alpha_invites, :beta_invites, :forever_account_value,
+        :permissioned_requests_served, :nonpermissioned_requests_served,
+        :can_post)
+    else
+      params.require(:person).permit(:email, :password, :password_confirmation, :remember_me)
+    end
+  end
+
 end
