@@ -1,63 +1,65 @@
-Privly::Application.configure do
-  # Settings specified here will take precedence over those in config/application.rb
+Rails.application.configure do
+  # Settings specified here will take precedence over those in config/application.rb.
 
   # The test environment is used exclusively to run your application's
-  # test suite.  You never need to work with it otherwise.  Remember that
+  # test suite. You never need to work with it otherwise. Remember that
   # your test database is "scratch space" for the test suite and is wiped
-  # and recreated between test runs.  Don't rely on the data there!
+  # and recreated between test runs. Don't rely on the data there!
   config.cache_classes = true
 
-  # Configure static asset server for tests with Cache-Control for performance
-  config.serve_static_assets = true
-  config.static_cache_control = "public, max-age=3600"
+  # Do not eager load code on boot. This avoids loading your whole application
+  # just for the purpose of running a single test. If you are using a tool that
+  # preloads Rails for running tests, you may have to set it to true.
+  config.eager_load = false
 
-  # Log error messages when you accidentally call methods on nil
-  config.whiny_nils = true
+  # Configure static file server for tests with Cache-Control for performance.
+  config.serve_static_files   = true
+  config.static_cache_control = 'public, max-age=3600'
 
-  # Show full error reports and disable caching
+  # Show full error reports and disable caching.
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
 
-  # Raise exceptions instead of rendering exception templates
+  # Raise exceptions instead of rendering exception templates.
   config.action_dispatch.show_exceptions = false
 
-  # Disable request forgery protection in test environment
-  config.action_controller.allow_forgery_protection    = false
+  # Disable request forgery protection in test environment.
+  config.action_controller.allow_forgery_protection = false
 
   # Tell Action Mailer not to deliver emails to the real world.
   # The :test delivery method accumulates sent emails in the
   # ActionMailer::Base.deliveries array.
   config.action_mailer.delivery_method = :test
 
-  # Use SQL instead of Active Record's schema dumper when creating the test database.
-  # This is necessary if your schema can't be completely dumped by the schema dumper,
-  # like if you have constraints or database-specific column types
-  # config.active_record.schema_format = :sql
+  # Randomize the order test cases are executed.
+  config.active_support.test_order = :random
 
-  # Print deprecation notices to the stderr
+  # Print deprecation notices to the stderr.
   config.active_support.deprecation = :stderr
 
-  # Allow pass debug_assets=true as a query parameter to load pages with unpackaged assets
-  config.assets.allow_debugging = true
-  
+  # Raises error for missing translations
+  # config.action_view.raise_on_missing_translations = true
+
+  # Where the email will come from
+  config.action_mailer.default_url_options = { :host => 'localhost:3000' }
+
   # Set this variable to change the displayed name of the site.
-  config.name = "test"
-  
+  config.name = "Site Name"
+
   # The host new injectible links should be created on.
-  # If you don't know what you are doing here, you should set
-  # it to the same domain as the primary_domain_host
-  config.link_domain_host = "test.host"
-  
+  # You should generally set this to the domain of your server
+  config.link_domain_host = "localhost:3000"
+
   config.action_mailer.default_url_options = { :host => 'localhost' }
-  
+
   # Required protocol, choose "http" or "https"
   config.required_protocol = "http"
-  
+
   # Don't send invitations to users by default.
   # Setting this to false will send users
   # accounts when they sign up for invitations,
   # while setting it to true will send them
   # a message that the system is currently in closed Alpha.
   config.send_invitations = false
-  
+
 end
