@@ -100,7 +100,7 @@ ActiveAdmin.setup do |config|
   # Admin comments are enabled by default.
   #
   # Default:
-  config.allow_comments = true
+  config.comments = true
   #
   # You can turn them on and off for any given namespace by using a
   # namespace config block.
