@@ -92,7 +92,7 @@ class ActiveAdminUserControllerTest < ActionController::TestCase
       :password_confirmation => "password"}
     assert_redirected_to admin_admin_user_path(:id => 1)
     
-    put :update, :id => 1
+    patch :update, :id => 1
     assert_redirected_to admin_admin_user_path(:id => 1)
     
     get :edit, :id => 1
