@@ -4,14 +4,16 @@ class NotifierTest < ActionMailer::TestCase
   test "Can send update email" do
     start_time = Time.now
     user = User.first
-    Notifier.update_invited_user(user)
+    mail = Notifier.update_invited_user(user)
+    mail.deliver_now
     assert user.last_emailed > start_time
   end
   
   test "Can send pending invitation email" do
     start_time = Time.now
     user = User.first
-    Notifier.pending_invitation(user)
+    mail = Notifier.pending_invitation(user)
+    mail.deliver_now
     assert user.last_emailed > start_time
   end
 end
